requirements.txt
aio-pika==9.3.0
aiormq==6.7.7
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beanie==1.22.6
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
lazy-model==0.2.0
motor==3.3.2
multidict==6.0.4
packaging==23.2
pamqp==3.2.1
passlib==1.7.4
pip-chill==1.0.3
pluggy==1.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.6.0
pytest==7.4.3
pytest-anyio==0.0.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1
rfc3986==2.0.0
sentry-sdk==1.31.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
azure-storage-blob==12.19.0
Jinja2==3.1.3
pycryptodome==3.20.0